ServiceDiscovery/requirements.txt
grpcio==1.59.0
grpcio-tools==1.41.0
protobuf==3.20.0
six==1.16.0
tenacity==8.2.3
